fdo#60789: writerfilter: do not insert a character at field start
The character inserted in DomainMapper_Impl::PushFieldContext() causes
problems in this bugdoc because there is a commentRangeStart immediately
before a text field and the call to delete it via setString("") disposes
the SwXTextRange that is stored in m_aAnnotationPosition.m_xStart,
which results in an exception when trying to insert the comment.

At least all the fields in the bugdoc import fine with this change.

Change-Id: I6c08d72fafb0529e8612f1e35b2c8b871edb5a52