fdo#59405 fdo#60638: EMFWriter::ImplWrite: write EMF_PLUS "comments"
When editing the Visio OLE object, there is a "preview" file generated,
which is apparently an EMF file (strangely initially inserting the Visio
object seems to result in a totally unproblematic WMF file).
The EMF file apparently has almost its entire content stored in
MetaCommentAction of type "EMF_PLUS", which is thrown away when writing
the file again.

Change-Id: I77a08454da673c1825aaa8421606737e7e8bc82c