Use __android_log_print() instead of fprintf(stderr)
Printing to stderr is not at all any faster or more direct in an Android app
than just using the Android standard logging API.

Note that in a normal Android app, stdout and stderr are not connected
anywhere. (Just like in "GUI" subsystem (as opposed to "console") Windows
programs, heh.) It is our own code in sal/android/lo-bootstrap.c that
redirects stdout and stderr to pipes which we set up and which are read in a
separate thread that we start. The lines read are then, surprise, passed on to
__android_log_print(). Thus writes to stdout or stderr in "normal" LibreOffice
code aren't lost.

But in code that is by definition Android-specific it makes little sense to
use stdout and stderr.

(Much of the affected logging in this change is in NativeActivity-related #if
0'ed code, sure, so won't actually be reached.)

Change-Id: I409114f36f3e535bb144b4bde0d378110b3336a1