make OutputDevice::ImplGetGraphics() return the SalGraphics*
Currently it returns int, meaning to be actually a bool returning
whether the call was successful. The pointer can work the same way,
but moreover allows access to SalGraphics* also from outside
of the class (otherwise I don't see much point in the method
being public).
To be used by code for embedded fonts handling for getting a list
of fonts etc.

Change-Id: I64f2acd61c1ce2f6c2e1e4aeac3f2af97c3198de