sort frame labels before frame contents for tab traversal
If there is more than one widget with the same mnemonic, and one is a frame
label, then we need to sort frame labels before frame bodies in the tab
traversal order. Otherwise if the focus is in the body of a frame whose label
has that shortcut and the shortcut is pressed again, the traversal through
following widgets will encounter the frame label as the next candidate, leading
back to the starting point and not onwards to the next widget using that
shortcut.

Frame labels have type "label" in the .ui, so suck that out to designate which
widget is the frame label at load time.

Change-Id: Ie21ed87867bd0c983981a3a8f3318b3cf598c1d6