Implement the VCLEVENT_WINDOW_ZOOM by turning it into a wheel mouse event
For now just use an ad-hoc factor that works close enough on my device to turn
the zoom scale factor into a value for the somewhat undocumented mnDelta field
in SalWheelMouseEvent. Ideally we should of course calculate it so that the
end result actually is the document content view being scaled by the requested
scale factor.

Change-Id: I21215d38c8ce4a35591d92d305ad1aeeca379f0a