fdo#56272 SwAnchoredDrawObject::_SetPositioningAttr: fix position
This reverts a part of 14f88b5088e9479a7e6b0bda233d20ec9efdea1f, that is
no longer needed, but causes incorrect drawing objects position.

The problem was that a non-zero position doesn't mean the layout in a
later iteration doesn't change the position, but the referred commit
didn't allow later iterations to fine-tune the position.

Change-Id: Id91a4c6952d1b58e0c5bc0b1a60991a284c29ddc