adapt code to define name .ui conversion
This is a bit of a pig, the RefButton does a magic
shrink of the dialog to just two widgets.

This happens in a baseclass which has other users
so for the time being switch off being in a
layout enabled dialog to do the right thing
depending on that.

On the bright side, for the other expandable
element we can now use a VclExpander and
throw away the custom resize, reposition, etc.
code for that.

Change-Id: I9f2ab8a8d1bd28ee873ec8548a26131eab597763