fdo#34800 Comments added to footer are placed at the right top of the page.
The problem is that the nodes in the Footnote and Footer are stored before
nodes of the document body in the internal structure.

Therefore I wrote this patch to check if the compared
comments are in the Footnote or Footer.
We don't need to check our comment is placed in the header
because it is already the most upper node of the whole document.

Test document: https://bugs.freedesktop.org/attachment.cgi?id=76038

Special thanks to Cédric Bosdonnat and Miklos Vajna

Change-Id: Ibcd0373110fde848dccf93ffe9100459c7cc64a5
Reviewed-on: https://gerrit.libreoffice.org/2572
Reviewed-by: Fridrich Strba <fridrich@documentfoundation.org>
Tested-by: Fridrich Strba <fridrich@documentfoundation.org>