SwTxtNode::GetRedlineTxt: fix string conversion
As I understand it, String allowed the length parameter to be longer
than the remaining string, in which case it simply returned the
substring till the end of the string. In case of OUString, "!!br0ken!!"
is returned instead, so make sure the length is never longer then the
remaining string.

Change-Id: I69499d74f1d9ba8ed6d9ba1a18d10b45ce319bf5