Don't crash the other experimental apps
Don't try to find the class org.libreoffice.experimental.desktop.Desktop in
the AndroidSalInstance constructor. It won't exist anyway except in that
specific app. Look up the class in the damaged() method where it is needed.

And actually, of course we should not hardcode the name of the app class like
that, but the app should pass its class down to the native code.

Change-Id: Ic15d5cc2c8d53be558711ca7a145d5489e34d298