Fix fallout from 63a911db0f5f8176023cdf67c488013c13db8939
WaE: expression which evaluates to zero treated as a null pointer
constant of type 'OutputDevice *'.

In this case, I think the code would have worked despite the scary
warning. (By luck, as the sal_False got passed as a NULL OutputDevice
pointer, and the following sal_Boolean parameter, which the sal_False
presumably was intended to mean, has the default value of sal_False.)
But better to make it more explicit which overload of Paint() is
intended.

But seriously, what good is supposed to come from the overloading of
the function name Paint() here? Would it really be so horribly
old-fashioned and C-like to use two different function names for the
slightly different use cases?

Change-Id: Ib397c8c1785320d3a9c0579a24f2c28ee4d22796