rhbz#915743: Avoid potentially blocking autocompletion for non file URLs
In theory, the logic of how SvtMatchContext_Impl threads are used to obtain
autocompletion information could probably be improved so that long-latency
scenarios would only block the ultimate closing of the dialog, if anything at
all.  In practice, this is probably not worth it, as users likely do not expect
an autocompletion feature for remote content, anyway.  (And previous commits
like 04079f599bbdedf436897b095871f75b6192b313 "neon commands cannot be aborted"
make it questionable whether any remaining blocking upon closing of the dialog
could be bounded in a reasonable manner.)

Change-Id: Ie82b9a70285b30fe88cf4c40992710087e86815e