fdo#60300: Work-in-progress change to rework pivot table core.
The idea is to avoid parsing the pivot table sheet output in order
to calculate GETPIVOTDATA.  The table outout is configurable, and it
will only be more configurable in the future.  The gist of my rework
is to calcualte the result of GETPIVOTDATA with the internl result
tree alone.

Also, the same result tree can be used for drill down too, which also
currently parses the table output, therefore subject to the same
limitation & fragility.

Change-Id: Ib0147e2aa2b710dfd627df7f535a685301214a52