CONTAINER_APPEND is SAL_ULONG, an unsigned long
Fix WaE: implicit conversion from 'unsigned long' to 'sal_uInt32' (aka
'unsigned int') changes value from 18446744073709551615 to 4294967295,
which definitely sounds like an actual bug-inducing error to me, not
"just a warning".

So don't try to use it as default value for a parameter of type
sal_uInt32. Changing the parameter to sal_uLong.

Change-Id: I8ebc7195ba1bf94f4f2d6df7cd7863f779ba3fa6