Check feature test macros instead of ANDROID and/or IOS
Pass -DLIBO_FEATURE_FOO to compiler for some elements in BUILD_TYPE.
BUILD_TYPE has at least two kinds of elements: Those that indicate
building a bundled copy of some 3rd-party library, and those that
indicate some specific feature of the LibreOffice platform or
build-time configuration choice. This is for the latter kind.

Change many of the checks for Android and/or iOS in the source code to
check LIBO_FEATURE_DESKTOP, LIBO_FEATURE_HELP or DISABLE_EXTENSIONS
instead, in cases where that is what is meant, not Android or iOS
specifically.

Change-Id: I2cd3f3bb99e953c7754dcea76a426f8f9d61e4db