Ifdef out methods for iOS that are actually in coretext/salgdi.cxx
A bit scary that the linker didn't notice the duplicate symbols, but
instead apparently just picked the first (or an arbitrary) one. Need
to look into that. Anyway, now when the correct SetTextColor is linked
in the coloured text in the demo document shows up in all its glory.

The code that is partially shared between the headless, OS X CoreText
and iOS backends should to be refactored a bit, I guess.

Change-Id: Id341298f72dc253380d9b2319032e0a9a8bdd0f6