fdo#58807 bring Beep() for basic back
The functionality was removed by fdo#48549.

This partially reverts

0069ad26968706d852122b7db515168ad3d3cafe
39f67064350d1f08d879d73285067f276ad5636a
2cb05101a68ff9852b258e1cd20e19b7c1e56e45

I changed the behavior to always beep, because only the basic macro
function is using Beep(). Looks like the Beep macro function didn't
even work correctly before the removal, because the default was to
not beep for most platforms. So I set the volume from disable (0)
to 50% for XBell().

Change-Id: I663ffb7af75d2fd6d2c1f94073e4412d9744de4a
Reviewed-on: https://gerrit.libreoffice.org/3124
Reviewed-by: Thorsten Behrens <tbehrens@suse.com>
Tested-by: Thorsten Behrens <tbehrens@suse.com>