fdo#32082: Do not reset number format in table cells on edit.
When we changed the default to _not_ to recognize the number formats in table
cells (commit eb26fe464b1d199eb74fac9c159828e0ce325f35), it caused that any
edit of a number formatted cell had reset it to text on edit.

This was quite unexpected, because when eg. there was a number that contributed
to a sum, and it changed to text, the sum changed as if there was a '0' there.

To solve this, try hard to retain the number format regardless of the
'Recognize Number Formats' setting when there was a number formatting assigned
to the cell.

Change-Id: Ib43e0877420d90cdac7a0357d1ef3a549c866fb7