Use SAL_OVERRIDE and drop an unused member function
Drop GetGlyphOutlines() which is not used, does not override any base
class function, and was even declared virtual even if there are no
derived classes.

Drop the dummy InitFont(), the equally dummy one in the base class is
good enough.

Also, put the overridden base class functions in the same order as in
the base class. Improves code readability and understandability.

Change-Id: Ia041ebaafae6c30cff18e00ee56f221e2f49b3e7